Refactor ZTunnel and Cni image digest under istiovalues digests (#2215)
Place ZTunnel and Cni image digests config apply under
"istiovalues/digests.go" where pilot image digest already reside.
