+++ pkg/istiovalues/digests_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	v1 "github.com/istio-ecosystem/sail-operator/api/v1"
 	"github.com/istio-ecosystem/sail-operator/pkg/config"
+	"github.com/stretchr/testify/assert"
 
 	"istio.io/istio/pkg/ptr"
 )
@@ -74,9 +75,6 @@ func TestApplyImageDigests(t *testing.T) {
 						Image: ptr.Of("proxy-test"),
 					},
 				},
-				// ZTunnel: &v1.ZTunnelConfig{
-				// 	Image: "ztunnel-test",
-				// },
 			},
 		},
 		{
@@ -108,9 +106,6 @@ func TestApplyImageDigests(t *testing.T) {
 						Image: ptr.Of("proxy-test"),
 					},
 				},
-				// ZTunnel: &v1.ZTunnelConfig{
-				// 	Image: "ztunnel-test",
-				// },
 			},
 		},
 		{
@@ -144,9 +139,6 @@ func TestApplyImageDigests(t *testing.T) {
 						Image: ptr.Of("proxy-test"),
 					},
 				},
-				// ZTunnel: &v1.ZTunnelConfig{
-				// 	Image: "ztunnel-test",
-				// },
 			},
 		},
 		{
@@ -227,6 +219,166 @@ func TestApplyImageDigests(t *testing.T) {
 			if diff := cmp.Diff(tc.expectValues, result); diff != "" {
 				t.Errorf("unexpected merge result; diff (-expected, +actual):\n%v", diff)
 			}
+		})
+	}
+}
+
+func TestApplyCNIImageDigests(t *testing.T) {
+	tests := []struct {
+		name     string
+		version  string
+		values   *v1.CNIValues
+		config   config.OperatorConfig
+		expected *v1.CNIValues
+	}{
+		{
+			name:    "no digests defined",
+			version: "v1.24.0",
+			values:  nil,
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{},
+			},
+			expected: nil,
+		},
+		{
+			name:    "applies digest when values is nil",
+			version: "v1.24.0",
+			values:  nil,
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {CNIImage: "istio/cni@sha256:abc123"},
+				},
+			},
+			expected: &v1.CNIValues{
+				Cni: &v1.CNIConfig{
+					Image: ptr.Of("istio/cni@sha256:abc123"),
+				},
+			},
+		},
+		{
+			name:    "does not override user-set global hub",
+			version: "v1.24.0",
+			values: &v1.CNIValues{
+				Global: &v1.CNIGlobalConfig{
+					Hub: ptr.Of("my-registry.io"),
+				},
+			},
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {CNIImage: "istio/cni@sha256:abc123"},
+				},
+			},
+			expected: &v1.CNIValues{
+				Global: &v1.CNIGlobalConfig{
+					Hub: ptr.Of("my-registry.io"),
+				},
+			},
+		},
+		{
+			name:    "does not override user-set image",
+			version: "v1.24.0",
+			values: &v1.CNIValues{
+				Cni: &v1.CNIConfig{
+					Image: ptr.Of("my-custom-image"),
+				},
+			},
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {CNIImage: "istio/cni@sha256:abc123"},
+				},
+			},
+			expected: &v1.CNIValues{
+				Cni: &v1.CNIConfig{
+					Image: ptr.Of("my-custom-image"),
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := ApplyCNIImageDigests(tt.version, tt.values, tt.config)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
+
+func TestApplyZTunnelImageDigests(t *testing.T) {
+	tests := []struct {
+		name     string
+		version  string
+		values   *v1.ZTunnelValues
+		config   config.OperatorConfig
+		expected *v1.ZTunnelValues
+	}{
+		{
+			name:    "no digests defined",
+			version: "v1.24.0",
+			values:  nil,
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{},
+			},
+			expected: nil,
+		},
+		{
+			name:    "applies digest when values is nil",
+			version: "v1.24.0",
+			values:  nil,
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
+				},
+			},
+			expected: &v1.ZTunnelValues{
+				ZTunnel: &v1.ZTunnelConfig{
+					Image: ptr.Of("istio/ztunnel@sha256:abc123"),
+				},
+			},
+		},
+		{
+			name:    "does not override user-set global hub",
+			version: "v1.24.0",
+			values: &v1.ZTunnelValues{
+				Global: &v1.ZTunnelGlobalConfig{
+					Hub: ptr.Of("my-registry.io"),
+				},
+			},
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
+				},
+			},
+			expected: &v1.ZTunnelValues{
+				Global: &v1.ZTunnelGlobalConfig{
+					Hub: ptr.Of("my-registry.io"),
+				},
+			},
+		},
+		{
+			name:    "does not override user-set image",
+			version: "v1.24.0",
+			values: &v1.ZTunnelValues{
+				ZTunnel: &v1.ZTunnelConfig{
+					Image: ptr.Of("my-custom-image"),
+				},
+			},
+			config: config.OperatorConfig{
+				ImageDigests: map[string]config.IstioImageConfig{
+					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
+				},
+			},
+			expected: &v1.ZTunnelValues{
+				ZTunnel: &v1.ZTunnelConfig{
+					Image: ptr.Of("my-custom-image"),
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := ApplyZTunnelImageDigests(tt.version, tt.values, tt.config)
+			assert.Equal(t, tt.expected, result)
 		})
 	}
 }
