+++ pkg/reconcile/ztunnel_test.go
@@ -18,16 +18,12 @@ import (
 	"context"
 	"testing"
 
-	v1 "github.com/istio-ecosystem/sail-operator/api/v1"
-	"github.com/istio-ecosystem/sail-operator/pkg/config"
 	"github.com/istio-ecosystem/sail-operator/pkg/reconciler"
 	"github.com/istio-ecosystem/sail-operator/pkg/scheme"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-
-	"istio.io/istio/pkg/ptr"
 )
 
 func TestZTunnelReconciler_Validate(t *testing.T) {
@@ -98,86 +94,6 @@ func TestZTunnelReconciler_Validate(t *testing.T) {
 			} else {
 				assert.NoError(t, err)
 			}
-		})
-	}
-}
-
-func TestApplyZTunnelImageDigests(t *testing.T) {
-	tests := []struct {
-		name     string
-		version  string
-		values   *v1.ZTunnelValues
-		config   config.OperatorConfig
-		expected *v1.ZTunnelValues
-	}{
-		{
-			name:    "no digests defined",
-			version: "v1.24.0",
-			values:  nil,
-			config: config.OperatorConfig{
-				ImageDigests: map[string]config.IstioImageConfig{},
-			},
-			expected: nil,
-		},
-		{
-			name:    "applies digest when values is nil",
-			version: "v1.24.0",
-			values:  nil,
-			config: config.OperatorConfig{
-				ImageDigests: map[string]config.IstioImageConfig{
-					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
-				},
-			},
-			expected: &v1.ZTunnelValues{
-				ZTunnel: &v1.ZTunnelConfig{
-					Image: ptr.Of("istio/ztunnel@sha256:abc123"),
-				},
-			},
-		},
-		{
-			name:    "does not override user-set global hub",
-			version: "v1.24.0",
-			values: &v1.ZTunnelValues{
-				Global: &v1.ZTunnelGlobalConfig{
-					Hub: ptr.Of("my-registry.io"),
-				},
-			},
-			config: config.OperatorConfig{
-				ImageDigests: map[string]config.IstioImageConfig{
-					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
-				},
-			},
-			expected: &v1.ZTunnelValues{
-				Global: &v1.ZTunnelGlobalConfig{
-					Hub: ptr.Of("my-registry.io"),
-				},
-			},
-		},
-		{
-			name:    "does not override user-set image",
-			version: "v1.24.0",
-			values: &v1.ZTunnelValues{
-				ZTunnel: &v1.ZTunnelConfig{
-					Image: ptr.Of("my-custom-image"),
-				},
-			},
-			config: config.OperatorConfig{
-				ImageDigests: map[string]config.IstioImageConfig{
-					"v1.24.0": {ZTunnelImage: "istio/ztunnel@sha256:abc123"},
-				},
-			},
-			expected: &v1.ZTunnelValues{
-				ZTunnel: &v1.ZTunnelConfig{
-					Image: ptr.Of("my-custom-image"),
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := ApplyZTunnelImageDigests(tt.version, tt.values, tt.config)
-			assert.Equal(t, tt.expected, result)
 		})
 	}
 }
